Move utility methods from GUIUtil to OfferViewUtil
